rename into IWorkerPool
